+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@ Things under "main" are not released yet.
 - Bring back ability to analyze flow types with `==` to a literal that is not the only value of that type.
 - Improve object flow typing with `std.length`.
 - Support `std.objectHasEx` in flow typing.
+- Remove what little support we had for named arguments in flow typing.
 
 ## v0.7.8