+++ docs/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 Things under "main" are not released yet.
 
-## main
+## v0.8.0
 
 - Bring back ability to analyze flow types with `==` to a literal that is not the only value of that type.
 - Improve object flow typing with `std.length`.